backend/requirements.txt
# Flask Backend Dependencies
flask==3.0.0
flask-cors==4.0.0

# Deep Learning
torch==2.1.0
torchvision==0.16.0

# YOLO-NAS (Super Gradients) - Required for blood cell detection
# Must match train_yolonas_blood_cells_optimized.py
super-gradients>=3.5.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Utilities
pathlib
pyyaml  # Required by Super Gradients
